refactor: move ecp into apps
