+++ .github/workflows/ci-it.yml
@@ -147,10 +147,9 @@ jobs:
       fail-fast: true
       matrix:
         package-paths:
-          - ./modules/apps/...
           - ./modules/cmp/...
 
-          - ./modules/dicehub/... ./modules/ecp/...
+          - ./modules/dicehub/...
 
 
           - ./modules/hepa/...
@@ -169,6 +168,7 @@ jobs:
           - ./modules/tools/monitor/...
           - ./modules/core/...
           - ./modules/apps/dop/...
+          - ./modules/apps/admin/... ./modules/apps/devflow/... ./modules/apps/ecp/... ./modules/apps/gallery/...
     steps:
       - name: Install Tools
         run: apk add tar
