+++ modules/apps/ecp/endpoints/endpoints.go
@@ -19,9 +19,9 @@ import (
 
 	clusterpb "github.com/erda-project/erda-proto-go/core/clustermanager/cluster/pb"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/modules/ecp/dbclient"
-	"github.com/erda-project/erda/modules/ecp/services/edge"
-	"github.com/erda-project/erda/modules/ecp/services/kubernetes"
+	"github.com/erda-project/erda/modules/apps/ecp/dbclient"
+	"github.com/erda-project/erda/modules/apps/ecp/services/edge"
+	"github.com/erda-project/erda/modules/apps/ecp/services/kubernetes"
 	"github.com/erda-project/erda/pkg/http/httpserver"
 )
